requirements.txt
h5py==3.7.0
matplotlib==3.5.2
mplfinance==0.12.9b1
numpy==1.23.1
pandas==1.4.3
pandas_ta==0.3.14b0
pybit==2.3.0
python_dateutil==2.8.2
